Add deprecation messages
